Added a helpers folder to better organize our things
