+++ helpers/drawing_helpers.py
