+++ helpers/qec_helpers.py
